+++ R/utils.R
@@ -517,6 +517,32 @@ config_netlify = function(output = 'netlify.toml', new_config = list()) {
   }
 }
 
+#' Create or modify the \file{.Rprofile} file for a website project
+#'
+#' If the file \file{.Rprofile} does not exist in the current directory, copy
+#' the file from the \file{resources} directory of \pkg{blogdown}. If the option
+#' \code{blogdown.hugo.version} is not found in this file, append
+#' \code{options(blogdown.hugo.version = "VERSION")} to it, where \code{VERSION}
+#' is obtained from \code{\link{hugo_version}()}.
+#' @export
+#' @return As a side-effect, the file \file{.Rprofile} is created or modified.
+config_Rprofile = function() {
+  f1 = '.Rprofile'; f2 = pkg_file('resources', f1)
+  if (file.exists(f1)) {
+    message("The file '", f1, "' exists, so I will not overwrite it with:")
+    message2(xfun::file_string(f2), files = f1)
+  }
+  file.copy(f2, f1, './', overwrite = FALSE)
+  ver = sprintf('\noptions(blogdown.hugo.version = "%s")\n', hugo_version())
+  if (!any(grepl('blogdown[.]hugo[.]version', read_utf8(f1)))) {
+    message2(
+      "I didn't find the option blogdown.hugo.version in '", f1,
+      "', so I will append the option to it.", files = f1
+    )
+    cat(ver, file = f1, append = TRUE)
+  }
+}
+
 # option names may be case insensitive
 get_config = function(field, default, config = load_config()) {
   config[[field]] %n% config[[match(tolower(field), tolower(names(config)))]] %n% default